Add files via upload
The networks are created in motion_head_3D_architectures_l3.py. "l3" is for linse3, but with the new keras version,they should work on linse9 as well. CNN's are trained in fTrain_fPredict_l3.py. archi_helpers_l3.py contains a few helper functions for the creation of the CNN's.